Sections 5 & 7-VertexAI-Development/Bikeshare-Model/Cloud-Functions-Batch-Prediction/requirements.txt
functions-framework>=3.0.0
google-cloud-aiplatform>=1.60.0
cloudevents>=1.9.0
